BUG: Pass integers to Bokeh RGB constructor (#1164)
* roudning error in RGB functionality

* Update backtesting/_plotting.py

+++ backtesting/_plotting.py
@@ -90,7 +90,7 @@ def colorgen():
 def lightness(color, lightness=.94):
     rgb = np.array([color.r, color.g, color.b]) / 255
     h, _, s = rgb_to_hls(*rgb)
-    rgb = np.array(hls_to_rgb(h, lightness, s)) * 255.
+    rgb = (np.array(hls_to_rgb(h, lightness, s)) * 255).astype(int)
     return RGB(*rgb)
